feat(animation): extract useAnimation hook
